+++ llama_index/llms/__init__.py
@@ -30,6 +30,7 @@ from llama_index.llms.localai import LOCALAI_DEFAULTS, LocalAI
 from llama_index.llms.mistral import MistralAI
 from llama_index.llms.mock import MockLLM
 from llama_index.llms.monsterapi import MonsterLLM
+from llama_index.llms.neutrino import Neutrino
 from llama_index.llms.nvidia_tensorrt import LocalTensorRTLLM
 from llama_index.llms.nvidia_triton import NvidiaTriton
 from llama_index.llms.ollama import Ollama
@@ -43,17 +44,6 @@ from llama_index.llms.portkey import Portkey
 from llama_index.llms.predibase import PredibaseLLM
 from llama_index.llms.replicate import Replicate
 from llama_index.llms.together import TogetherLLM
-from llama_index.llms.types import (
-    ChatMessage,
-    ChatResponse,
-    ChatResponseAsyncGen,
-    ChatResponseGen,
-    CompletionResponse,
-    CompletionResponseAsyncGen,
-    CompletionResponseGen,
-    LLMMetadata,
-    MessageRole,
-)
 from llama_index.llms.vertex import Vertex
 from llama_index.llms.vllm import Vllm, VllmServer
 from llama_index.llms.watsonx import WatsonX
@@ -93,6 +83,7 @@ __all__ = [
     "MessageRole",
     "MockLLM",
     "MonsterLLM",
+    "Neutrino",
     "NvidiaTriton",
     "MistralAI",
     "Ollama",