+++ llama_index/query_pipeline/__init__.py
@@ -8,7 +8,12 @@ from llama_index.core.query_pipeline.components import (
 )
 from llama_index.core.query_pipeline.query_component import (
     CustomQueryComponent,
+    Link,
     QueryComponent,
+)
+from llama_index.query_pipeline.components.router import (
+    RouterComponent,
+    SelectorComponent,
 )
 from llama_index.query_pipeline.query import InputKeys, OutputKeys, QueryPipeline
 
@@ -22,4 +27,7 @@ __all__ = [
     "FnComponent",
     "ArgPackComponent",
     "KwargPackComponent",
+    "RouterComponent",
+    "SelectorComponent",
+    "Link",
 ]