+++ llama_index/question_gen/output_parser.py
@@ -12,6 +12,12 @@ class SubQuestionOutputParser(BaseOutputParser):
         if not json_dict:
             raise ValueError(f"No valid JSON found in output: {output}")
 
+        # example code includes an 'items' key, which breaks
+        # the parsing from open-source LLMs such as Zephyr.
+        # This gets the actual subquestions and recommended tools directly
+        if "items" in json_dict:
+            json_dict = json_dict["items"]
+
         sub_questions = [SubQuestion.parse_obj(item) for item in json_dict]
         return StructuredOutput(raw_output=output, parsed_output=sub_questions)