+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # ChangeLog
 
+## [0.9.33] - 2024-01-17
+
+### New Features
+
+- Added RankGPT as a postprocessor (#10054)
+- Ensure backwards compatibility with new Pinecone client version bifucation (#9995)
+- Recursive retriever all the things (#10019)
+
+### Bug Fixes / Nits
+
+- BugFix: When using markdown element parser on a table containing comma (#9926)
+- extend auto-retrieval notebook (#10065)
+- Updated the Attribute name in llm_generators (#10070)
+- jaguar vector store add text_tag to add_kwargs in add() (#10057)
+
+## [0.9.32] - 2024-01-16
+
+### New Features
+
+- added query-time row retrieval + fix nits with query pipeline over structured data (#10061)
+- ReActive Agents w/ Context + updated stale link (#10058)
+
+## [0.9.31] - 2024-01-15
+
+### New Features
+
+- Added selectors and routers to query pipeline (#9979)
+- Added sparse-only search to qdrant vector store (#10041)
+- Added Tonic evaluators (#10000)
+- Adding async support to firestore docstore (#9983)
+- Implement mongodb docstore `put_all` method (#10014)
+
+### Bug Fixes / Nits
+
+- Properly truncate sql results based on `max_string_length` (#10015)
+- Fixed `node.resolve_image()` for base64 strings (#10026)
+- Fixed cohere system prompt role (#10020)
+- Remove redundant token counting operation in SentenceSplitter (#10053)
+
 ## [0.9.30] - 2024-01-11
 
 ### New Features